feat: implement more copilot changes
